Fixing controller tests after renaming
